Organize files into modular file structure
